# swagger.yaml
openapi: 3.0.2
info:
  title: Railroad API
  description: API documentation for Railroad, a train app
  version: 1.0.0
servers:
  - url: http://localhost:3000/
paths:
...
  /users:
    get:
      summary: Get a list of users
      description: Requires employee privileges and a valid token
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      security:
        - BearerAuth: [ ]
  /users/register:
    post:
      summary: Register a new user
      description: Creates a new user with the provided credentials
      requestBody:
        description: User registration object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterPayload'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/login:
    post:
      summary: Login with user credentials
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginPayload'
      responses:
        '200':
          description: User logged in successfully
        '401':
          description: Unauthorized
  /users/{id}:
    get:
      summary: Get details of a specific user
      description: Requires a valid token and checks if the user exists
      parameters:
        - name: id
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
      security:
        - BearerAuth: [ ]
    put:
      summary: Update a specific user
      description: Requires a valid token and checks user privileges
      parameters:
        - name: id
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
      requestBody:
        description: Updated user object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdatePayload'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: User privileges not sufficient
        '404':
          description: User not found
      security:
        - BearerAuth: [ ]
    delete:
      summary: Delete a specific user
      description: Requires a valid token and checks user privileges
      parameters:
        - name: id
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '403':
          description: User privileges not sufficient
        '404':
          description: User not found
      security:
        - BearerAuth: [ ]
...
components:
  schemas:
...
    User:
      type: object
      properties:
        email:
          type: string
          description: Email of the user
        role:
          type: string
          enum: [ 'Admin', 'Employee', 'User' ]
          description: Role of the user
        lastEdited:
          type: string
          format: date-time
          description: Last edited timestamp
      required:
        - email
        - role
    UserRegisterPayload:
      type: object
      properties:
        username:
          type: string
          description: Username of the user
        email:
          type: string
          format: email
          description: Email of the user
        password:
          type: string
          description: "Password of the user (minimum length: 6)"
      required:
        - username
        - email
        - password
    UserUpdatePayload:
      type: object
      properties:
        username:
          type: string
          description: Updated username of the user
        email:
          type: string
          format: email
          description: Updated email of the user
        password:
          type: string
          description: Updated password of the user
      required:
        - username
        - email
        - password
    UserLoginPayload:
      type: object
      properties:
        username:
          type: string
          description: Username of the user
        password:
          type: string
          description: Password of the user
      required:
        - username
        - password
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: cookie
      name: token
      description: Use a valid bearer token for authentication